+++ classes/task/purge_cache.php
@@ -1,0 +1,61 @@
+<?php
+// This file is part of Moodle - http://moodle.org/
+//
+// Moodle is free software: you can redistribute it and/or modify
+// it under the terms of the GNU General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// Moodle is distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+// GNU General Public License for more details.
+//
+// You should have received a copy of the GNU General Public License
+// along with Moodle.  If not, see <http://www.gnu.org/licenses/>.
+
+/**
+ * Theme Boost Union - Scheduled task for purging the cache.
+ *
+ * @package    theme_boost_union
+ * @copyright  2024 Alexander Bias, lern.link GmbH <alexander.bias@lernlink.de>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+
+namespace theme_boost_union\task;
+
+defined('MOODLE_INTERNAL') || die();
+
+require_once(__DIR__ . '/../../locallib.php');
+
+/**
+ * The theme_boost_union purge_cache task class.
+ *
+ * @package    theme_boost_union
+ * @copyright  2024 Alexander Bias, lern.link GmbH <alexander.bias@lernlink.de>
+ * @license    http://www.gnu.org/copyleft/gpl.html GNU GPL v3 or later
+ */
+class purge_cache extends \core\task\scheduled_task {
+
+    /**
+     * Return localised task name.
+     *
+     * @return string
+     */
+    public function get_name() {
+        return get_string('task_purgecache', 'theme_boost_union');
+    }
+
+    /**
+     * Execute scheduled task.
+     *
+     * @return boolean
+     */
+    public function execute() {
+        // Simply reset the theme cache.
+        theme_reset_all_caches();
+
+        // Return true, just to keep the Moodle scheduler happy.
+        return true;
+    }
+}